@@ -21,33 +21,24 @@
 struct hkRootLevelContainer;
 
 struct HK_EXTERN IhkPackFile {
-  typedef std::unique_ptr<IhkVirtualClass> VirtualClass;
-  typedef std::vector<VirtualClass> VirtualClasses;
-  typedef std::vector<IhkVirtualClass *> VirtualClassesRef;
+  using VirtualClass = std::unique_ptr<IhkVirtualClass>;
+  using VirtualClasses = std::vector<VirtualClass>;
+  using VirtualClassesRef = std::vector<IhkVirtualClass *>;
+  using Ptr = std::unique_ptr<IhkPackFile>;
 
   virtual VirtualClasses &GetAllClasses() = 0;
   virtual hkToolset GetToolset() const = 0;
-  virtual ~IhkPackFile(){};
-
-  VirtualClassesRef GetClasses(es::string_view hkClassName) {
-    return GetClasses(JenkinsHash(hkClassName));
-  }
+  virtual ~IhkPackFile() = default;
 
   VirtualClassesRef GetClasses(JenHash hash);
   hkRootLevelContainer *GetRootLevelContainer();
   const IhkVirtualClass *GetClass(const void *ptr);
 
-  static IhkPackFile *Create(const std::string &fileName,
-                             bool suppressErrors = false);
-  static IhkPackFile *Create(es::string_view fileName,
-                             bool suppressErrors = false);
+  static Ptr Create(const std::string &fileName);
 
-  int ToXML(es::string_view fileName, hkToolset toolset);
-  int ToXML(const std::string &fileName, hkToolset toolset);
-
-  pugi::xml_document ToXML(hkToolset toolset);
+  void ToXML(const std::string &fileName, hkToolset toolset);
+  void ToXML(pugi::xml_node node, hkToolset toolset);
 
   // rule must be as hex, ie. 0x4101
-  int ToPackFile(es::string_view fileName, hkToolset toolset, uint32 rule);
-  int ToPackFile(const std::string &fileName, hkToolset toolset, uint32 rule);
+  void ToPackFile(const std::string &fileName, hkToolset toolset, uint32 rule);
 };
