@@ -1,10 +1,10 @@
 #pragma once
 #include "datas/pythonex.hpp"
 #include "datas/jenkinshash.hpp"
-#include <unordered_map>
+#include <map>
 
 struct IhkVirtualClass;
 
-typedef std::unordered_map<JenHash, PyObject *(*)(const IhkVirtualClass *)>
+typedef std::map<JenHash, PyObject *(*)(const IhkVirtualClass *)>
     HKPYRegistry;
 extern const HKPYRegistry hkClasses;
